LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/textfieldselection/selection-start-end-extra-expected.txt


PASS Setting defaultValue in a textarea should NOT move the cursor to the end
FAIL Setting defaultValue in a textarea with a value should NOT make any difference assert_equals: expected 6 but got 0
PASS Setting textContent in a textarea should NOT move selection{Start,End} to the end
PASS Adding children to a textarea should NOT move selection{Start,End} to the end
PASS Removing children from a textarea should NOT update selection{Start,End}
PASS Setting the same value (with different newlines) in a textarea should NOT update selection{Start,End}
FAIL Removing child nodes in non-dirty textarea should make selection{Start,End} 0 assert_equals: selectionStart after node removal expected 0 but got 3
FAIL Setting value to a shorter string than defaultValue should correct the cursor position assert_equals: expected 2 but got 3
PASS Shortening value by turning the input type into 'url' should correct selection{Start,End}
FAIL Shortening value by turning the input type into 'color' and back to 'text' should correct selection{Start,End} assert_equals: expected 0 but got 9
PASS Resetting a value to a shorter string than defaultValue should correct the cursor position
